+++ app/Http/Controllers/GenerosController.php
@@ -27,28 +27,44 @@ class GenerosController extends Controller
     }
     
      public function create(){
-        return view('generos.create');
+        if (Gate::allows('admin')){
+         return view('generos.create');
+            }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
     }
         public function store(Request $request){
-        $novoGenero = $request->validate([
+        if (Gate::allows('admin')){
+            $novoGenero = $request->validate([
         'designacao'=>['required', 'min:3', 'max:30'],
         'observacoes'=>['required', 'min:3', 'max:255'],
         ]);
          $genero = Genero::create($novoGenero);
         return redirect()->route('editoras.show', [
             'id'=>$genero->id_genero]);
+            }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
     }
     
     public function edit (Request $request){
+        if (Gate::allows('admin')){
         $id = $request->id;
         $genero = genero::where('id_genero',$id)->first();
         //dd ($genero);
         return view('generos.edit',[
             'genero'=>$genero
         ]);
+             }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
     }
     
     public function update (Request $request){
+        if (Gate::allows('admin')){
         $id = $request->all();
         $genero = Genero::findOrFail ($id);
         $updateGenero = $request->validate([
@@ -59,9 +75,14 @@ class GenerosController extends Controller
         return redirect()->route('generos.show', [ 
         'id'=>$genero->id_genero
         ]);
+            }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
     }
     
     public function delete (Request $request){
+         if (Gate::allows('admin')){
         $genero = Genero::where ('id_genero', $request->id )->first();
         if(is_null($genero)){
             return redirect()->route('generos.index')->with('mensagem','o genero nao existe');
@@ -70,11 +91,20 @@ class GenerosController extends Controller
         {
             return view('generos.delete',['genero'=>$genero]);
         }
+             }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
     }
     public function destroy(Request $request){
+        if (Gate::allows('admin')){
         $idGenero = $request->id;
         $genero = Genero::findOrFail($idGenero);
         $genero->delete();
         return redirect()->route('generos.index')->with('mensagem','genero eleminado!');
-    }
+        }
+        else{
+            return redirect()->route('livros.index')->with('mensagem','Nao tem permissão para aceder a area pretendida');
+        }
+        }        
 }
